Move fetchPageNumbers to utils
