+++ packages/Pagination/src/Paginator/base.tsx
@@ -3,7 +3,7 @@
  */
 import React, { Fragment, MouseEvent, useState } from 'react';
 import { Pagination, PaginationItem, PaginationLink } from 'reactstrap';
-import { range } from './utils';
+import { fetchPageNumbers } from './utils';
 
 /** interface for argument passed to the opPageChange prop */
 export interface PaginationData {
@@ -48,7 +48,7 @@ const Paginator = (props: PaginatorProps) => {
 
   /** click Handler; registers the number of the clicked pagination Item
    * handles click event
-   * @param {number} - number of the clicked plaginationItem
+   * @param {number} - number of the clicked paginationItem
    * @return {(e: MouseEvent) => void }
    */
   const handleClick = (page: number) => (e: MouseEvent) => {
@@ -151,36 +151,6 @@ const Paginator = (props: PaginatorProps) => {
       </Pagination>
     </Fragment>
   );
-};
-
-// helper && utilities
-
-/** Let's say we have 10 pages and we set neighbourPillsNum to 2
- * Given that the current page is 6
- * The pagination control will look like the following:
- *
- * (start) previous {4 5} [6] {7 8} next (end)
- *
- * (x) => terminal pages: first and last page(always visible)
- * [x] => represents current page
- * {...x} => represents page neighbours
- *
- * @param {number} neighbourPillsNum - the max number of links/pills on either side of current page
- * @param {number} totalPages - the total number of pages
- * @param {number} currentPage - the currently selected page
- * @return {number []} - Array of numbers that will form the pills between previous and next
- */
-export const fetchPageNumbers = (
-  neighbourPillsNum: number,
-  totalPages: number,
-  currentPage: number
-) => {
-  const startPage = Math.max(1, currentPage - neighbourPillsNum);
-  const endPage = Math.min(totalPages + 1, currentPage + neighbourPillsNum + 1);
-
-  const numberedPages: number[] = range(startPage, endPage);
-
-  return numberedPages;
 };
 
 Paginator.defaultProps = defaultPaginatorProps;
