+++ packages/Pagination/src/Paginator/utils.ts
@@ -18,3 +18,31 @@ export const range = (start: number, stop: number | null = null, step: number = 
     .fill(begin)
     .map((x, y) => x + y * step);
 };
+
+/** Let's say we have 10 pages and we set neighbourPillsNum to 2
+ * Given that the current page is 6
+ * The pagination control will look like the following:
+ *
+ * (start) previous {4 5} [6] {7 8} next (end)
+ *
+ * (x) => terminal pages: first and last page(always visible)
+ * [x] => represents current page
+ * {...x} => represents page neighbours
+ *
+ * @param {number} neighbourPillsNum - the max number of links/pills on either side of current page
+ * @param {number} totalPages - the total number of pages
+ * @param {number} currentPage - the currently selected page
+ * @return {number []} - Array of numbers that will form the pills between previous and next
+ */
+export const fetchPageNumbers = (
+  neighbourPillsNum: number,
+  totalPages: number,
+  currentPage: number
+) => {
+  const startPage = Math.max(1, currentPage - neighbourPillsNum);
+  const endPage = Math.min(totalPages + 1, currentPage + neighbourPillsNum + 1);
+
+  const numberedPages: number[] = range(startPage, endPage);
+
+  return numberedPages;
+};
